upgrade rxjs and typescript versions
after typescript 2.4.2, there introduced an error that was fixed in rxjs 5.4.3. see the error and issue here: https://github.com/ReactiveX/rxjs/issues/2705

ive upgraded the dependencies hoping to resolve this error. this repo is obsolte as of a couple months ago when this issue first arose. it will remain deprecated until you upgrade the versions. Hope you see this and accept my PR.